Updaed Level 2
At the end of one platform I added a level door. I want to code it to transport the main character to a floating island up above where there is a big boss waiting and they can collect something in a house (instead of it transporting the character to Nadja's level). Also, there are fireflies on the last platform that will transport the character to the next level. But this all has to be coded. Please let me know if there are any other changes I need to apply. I tried to make it as cohesive to the previous levels as possible.